Add array/slice support for reflected values
This adds support for using `AddReflected` or `AppendReflected` with an
array or slice.

It also resolves an issue where using `AddReflected` would invoke
`Elem()` on the reflected value which was an illegal usage of this
function and would cause a panic. If it hadn't caused a panic, it would
have likely resulted in an infinite loop anyway.

This also refactors the code slightly so `AddReflected` and
`AppendReflected` no longer have duplicate code and `AppendReflected` is
the primary section of code.